ISSUE #5982 - Refactored ticketsgroup; moved into its own file renamed for clarity
