+++ frontend/src/v5/ui/routes/dashboard/projects/tickets/tabularView/tabularViewTables/tabularViewTables.component.tsx
@@ -21,7 +21,7 @@ import { EmptyPageView } from '../../../../../../components/shared/emptyPageView
 import { ResizableTableContext } from '@controls/resizableTableContext/resizableTableContext';
 import { TicketsHooksSelectors } from '@/v5/services/selectorsHooks';
 import { templateAlreadyFetched } from '@/v5/store/tickets/tickets.helpers';
-import { GroupedTables, TicketsTableResizableContentProps } from './groupedTables/groupedTables.component';
+import { GroupedTables, GroupedTablesProps } from './groupedTables/groupedTables.component';
 import { Container, TicketsTableSpinner } from './tabularViewTables.styles';
 import { useEdgeScrolling } from '../edgeScrolling';
 import { BaseProperties } from '@/v5/ui/routes/viewer/tickets/tickets.constants';
@@ -54,7 +54,7 @@ const NoTicketsMessage = () => (
 	</Container>
 );
 
-export const TabularViewTables = ({ template, hasTickets, ...props }: TicketsTableResizableContentProps & { hasTickets: boolean }) => {
+export const TabularViewTables = ({ template, hasTickets, ...props }: GroupedTablesProps & { hasTickets: boolean }) => {
 	const tableRef = useRef(null);
 	const edgeScrolling = useEdgeScrolling();
 	const defaultColumns = TicketsHooksSelectors.selectInitialTabularColumns(template._id);
